cleaned matchHistory, stats_matchs, split to crad, keep prev
